+++ Irvin.Fludal/Irvin.Fludal.csproj
@@ -15,4 +15,12 @@
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     </PropertyGroup>
 
+    <ItemGroup>
+        <ProjectReference Include="..\..\system-extensions\SystemExtensions.Standard\SystemExtensions.Standard.csproj" Condition=" '$(Configuration)' == 'DebugEverything' " />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Irvin.SystemExtensions" Version="1.3.1" Condition=" '$(Configuration)' != 'DebugEverything' " />
+    </ItemGroup>
+
 </Project>
